requirements.txt
blinker==1.4
cffi==1.14.5
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-User==1.0.2.2
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
passlib==1.7.4
pycparser==2.20
pymongo==3.11.4
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
python-dotenv==0.18.0
cryptography==3.4.7
Babel==2.9.1
pytz==2021.1
Flask-Babel==2.0.0
Pillow==8.3.1
Flask-Dance==5.0.0
requests==2.26.0
charset-normalizer==2.0.4
certifi==2021.5.30
idna==3.2
oauthlib==3.1.1
requests-oauthlib==1.3.0
urllib3==1.26.6
URLObject==2.4.3
bidict=0.21.3
eventlet=0.32.0
flask-socketio=5.1.1
python-engineio=4.2.1
python-socketio=5.4.0
